+++ api/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Added `(GET) /openapi` which returns the OpenAPI specification of the API
 - Added `(GET) /changelog` which returns this changelog
+- Added `(GET) /ventures` which handles create, read, update, and delete actions for `Ventures`
 - Added `(GET) /things` which returns Things or a Thing
 - Updated `(GET) /things` with the `id` query parameter so a specific Thing can be returned
 - Updated `(GET) /things` to return Things or a Thing with the properties `description`, `id`, `child_ids`, `parent_ids`, `state`, `additional`, and `is_dead`